requirements.txt
Django>=5.0.6
djangorestframework>=3.15.2
djangorestframework-jwt>=1.11.0
pyjwt>=2.1.0